+++ src/main/java/nl/enjarai/showmeyourskin/mixin/ArmorFeatureRendererMixin.java
@@ -13,7 +13,9 @@ import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ArmorItem;
 import net.minecraft.util.Identifier;
+import net.minecraft.util.math.MathHelper;
 import nl.enjarai.showmeyourskin.config.ModConfig;
+import nl.enjarai.showmeyourskin.util.CombatLogger;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -57,11 +59,12 @@ public abstract class ArmorFeatureRendererMixin<T extends LivingEntity, M extend
     )
     private void armorTransparency(MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, ArmorItem item, boolean usesSecondLayer, A model, boolean legs, float red, float green, float blue, String overlay, CallbackInfo ci) {
         if (entity instanceof PlayerEntity player) {
-            var t = ModConfig.INSTANCE.getApplicable(player.getUuid()).getTransparency(slot);
-            if (t < 100) {
+            var t = ModConfig.INSTANCE.getApplicableTransparency(player.getUuid(), slot);
+
+            if (t < 1) {
                 if (t > 0) {
                     VertexConsumer vertexConsumer = ItemRenderer.getDirectItemGlintConsumer(vertexConsumers, RenderLayer.getEntityTranslucent(getArmorTexture(item, legs, overlay)), false, usesSecondLayer);
-                    model.render(matrices, vertexConsumer, light, OverlayTexture.DEFAULT_UV, red, green, blue, t / 100f);
+                    model.render(matrices, vertexConsumer, light, OverlayTexture.DEFAULT_UV, red, green, blue, t);
                 }
 
                 ci.cancel();